Hackathon - API/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
pyodbc==4.0.30
azure-ai-textanalytics==5.3.0